requirements.txt
# Heart Disease MLOps Project - Requirements
# Python 3.11+

# ============================================
# Core ML Libraries
# ============================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ============================================
# Experiment Tracking
# ============================================
mlflow>=2.9.0

# ============================================
# API Framework
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ============================================
# Data Visualization (for EDA)
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ============================================
# Jupyter Notebooks
# ============================================
jupyter>=1.0.0
ipykernel>=6.27.0
nbformat>=5.9.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# ============================================
# Code Quality
# ============================================
ruff>=0.1.6
black>=23.11.0

# ============================================
# Monitoring & Metrics
# ============================================
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# ============================================
# Model Serialization
# ============================================
joblib>=1.3.0

# ============================================
# Configuration
# ============================================
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ============================================
# HTTP Requests (for data download)
# ============================================
requests>=2.31.0

